+++ src/main/java/com/drawing/springboot/dto/CategoryDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class CategoryDTO {
-	private Long categoryId;  // CATEGORY_ID
+	private int categoryId;  // CATEGORY_ID
     private String name;      // NAME
     private String url;       // URL
     private String image; 	  // 이미지 url
